Split hia2reports from EC Repository
