+++ opensrp-app/src/main/java/org/smartregister/repository/BaseRepository.java
@@ -2,6 +2,8 @@ package org.smartregister.repository;
 
 import net.sqlcipher.database.SQLiteDatabase;
 
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
 import java.util.UUID;
 
 
@@ -15,6 +17,9 @@ public class BaseRepository {
     public static String TYPE_Synced = "Synced";
     public static String TYPE_Valid = "Valid";
     public static String TYPE_InValid = "Invalid";
+    protected static final String ORDER_BY = " order by ";
+
+    public static DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
 
     public static String COLLATE_NOCASE = " COLLATE NOCASE ";
